requirements.txt
Flask==3.0.3
google_api_python_client==2.125.0
google_auth_oauthlib==0.4.6
matplotlib==3.8.4
numpy==1.23.0
pandas==2.2.1
protobuf==3.12.4
python_dateutil==2.8.2
python_dateutil==2.8.1
scikit_learn==1.3.2
scipy==1.13.0
seaborn==0.13.2
